Fix all audit-identified bugs across the codebase
Critical:
- LocalModelManager: replace runBlocking with a @Volatile cache maintained
  by a coroutine observer, eliminating the potential main-thread block
- CustomLocalModelProvider: use suspend getActiveModel() in suspend fns

High:
- TrackingService: add 5-min withTimeoutOrNull to GPS-suspension wait so
  the loop cannot hang indefinitely if motion sensors go silent
- TrackingService: re-throw CancellationException; on unexpected Exception
  ensure the active track is ended so DB is never left in an open state
- TrackingService/BatteryMonitor: fix integer overflow in battery percent
  calculation by using toLong() before multiplying

Medium:
- DashboardViewModel: debounce(500ms) stats recomputation to prevent
  redundant queries on every point insert during active recording
- TrackRepository: migrate Geocoder to the non-deprecated callback API
  on Android 13+ (TIRAMISU); keep legacy path for older devices
- SettingsDataStore: change webhook URL fallback from hardcoded domain
  to empty string so unconfigured state is unambiguous

Low / Safety:
- GpsSatelliteTracker: capture callback in local val before registration,
  eliminating the !! force-unwrap
- WearableManager: guard STATE_CONNECTED against null currentDevice;
  capture lastCrankRevs/lastCrankTime to local vals inside null check
- SettingsScreen: extract spO2 and temperatureC to non-null local vals
  instead of repeated !! dereferences inside null guards
- TracksScreen: capture exportTrackId to local val to eliminate !!
- TrackDatabase: narrow type-converter catch to IllegalArgumentException
  instead of catching all Throwable subclasses
- LocalModelManager: same narrowing for ModelInstallState converter

+++ app/src/main/java/com/trackjourney/data/location/BatteryMonitor.kt
@@ -61,7 +61,7 @@ class BatteryMonitor(private val context: Context) {
 
         val level = intent.getIntExtra(BatteryManager.EXTRA_LEVEL, -1)
         val scale = intent.getIntExtra(BatteryManager.EXTRA_SCALE, -1)
-        val percent = if (level >= 0 && scale > 0) (level * 100) / scale else -1
+        val percent = if (level >= 0 && scale > 0) (level.toLong() * 100 / scale).toInt() else -1
 
         _batteryState.value = BatteryState(isCharging = isCharging, level = percent)
     }
